Removed class which now has functionality duplicated in Weather file
